lab3/requirements-airflow.txt
apache-airflow
pandas
s3fs
boto3
scikit-learn
python-dotenv